Tighten Python SDK cleanup
