+++ sdk/python/gestalt/_pyinstaller.py
@@ -1,5 +1,4 @@
 from gestalt._runtime import main
 
-
 if __name__ == "__main__":
     raise SystemExit(main())
