+++ sdk/python/scripts/generate_stubs.py
@@ -6,7 +6,6 @@ import sys
 import tempfile
 from pathlib import Path
 
-
 EXPECTED_GRPC_IMPORT = "from v1 import plugin_pb2 as v1_dot_plugin__pb2\n"
 RELATIVE_GRPC_IMPORT = "from . import plugin_pb2 as v1_dot_plugin__pb2\n"
 GRPC_RUNTIME_HEADER = """import grpc
